preparing for test 42

+++ lib_ft/ft_memchr.c
@@ -1,7 +1,6 @@
-#include <string.h>
-#include <stdio.h>
+#include "libft.h"
 
-void *ft_memchr(const void *s, int c, unsigned int n)
+void *ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char *src;
